requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
channels==4.0.0
channels-redis==4.1.0
Django==5.0.1
djangorestframework==3.14.0
msgpack==1.0.7
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
tzdata==2023.4
